+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/preview/AutocraftingPreviewListener.java
@@ -10,4 +10,6 @@ interface AutocraftingPreviewListener {
     void previewChanged(@Nullable Preview preview);
 
     void requestRemoved(AutocraftingRequest request, boolean last);
+
+    void maxAmountReceived(double maxAmount);
 }