+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/preview/AutocraftingPreviewScreen.java
@@ -3,11 +3,14 @@ package com.refinedmods.refinedstorage.common.autocrafting.preview;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.Preview;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewItem;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewType;
+import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
+import com.refinedmods.refinedstorage.common.Platform;
 import com.refinedmods.refinedstorage.common.api.RefinedStorageClientApi;
 import com.refinedmods.refinedstorage.common.api.support.resource.ResourceRendering;
 import com.refinedmods.refinedstorage.common.support.amount.AbstractAmountScreen;
 import com.refinedmods.refinedstorage.common.support.amount.AmountScreenConfiguration;
 import com.refinedmods.refinedstorage.common.support.amount.DoubleAmountOperations;
+import com.refinedmods.refinedstorage.common.support.tooltip.HelpClientTooltipComponent;
 import com.refinedmods.refinedstorage.common.support.tooltip.SmallText;
 import com.refinedmods.refinedstorage.common.support.widget.ScrollbarWidget;
 
@@ -17,11 +20,13 @@ import javax.annotation.Nullable;
 
 import com.google.common.util.concurrent.RateLimiter;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.renderer.Rect2i;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.Style;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 import org.joml.Vector3f;
@@ -38,9 +43,31 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
     private static final MutableComponent TITLE = createTranslation("gui", "autocrafting_preview.title");
     private static final MutableComponent START = createTranslation("gui", "autocrafting_preview.start");
     private static final MutableComponent PENDING = createTranslation("gui", "autocrafting_preview.pending");
+    private static final MutableComponent MAX = createTranslation("gui", "autocrafting_preview.max");
+    private static final MutableComponent MAX_HELP = createTranslation("gui", "autocrafting_preview.max.help");
     private static final MutableComponent MISSING_RESOURCES = createTranslation(
         "gui",
         "autocrafting_preview.start.missing_resources"
+    );
+    private static final MutableComponent CYCLE_DETECTED = createTranslation(
+        "gui",
+        "autocrafting_preview.cycle_detected"
+    ).withStyle(Style.EMPTY.withBold(true));
+    private static final MutableComponent CYCLE_OUTPUTS = createTranslation(
+        "gui",
+        "autocrafting_preview.cycle_detected.outputs"
+    );
+    private static final MutableComponent BREAK_THE_CYCLE_AND_TRY_AGAIN = createTranslation(
+        "gui",
+        "autocrafting_preview.cycle_detected.break_the_cycle_and_try_again"
+    );
+    private static final MutableComponent REQUEST_TOO_LARGE_TO_HANDLE = createTranslation(
+        "gui",
+        "autocrafting_preview.request_too_large_to_handle"
+    ).withStyle(Style.EMPTY.withBold(true));
+    private static final MutableComponent TRY_SMALLER_AMOUNT = createTranslation(
+        "gui",
+        "autocrafting_preview.request_too_large_to_handle.try_smaller_amount"
     );
     private static final ResourceLocation ROW = createIdentifier("autocrafting_preview/row");
     private static final ResourceLocation CRAFTING_REQUESTS = createIdentifier("autocrafting_preview/requests");
@@ -62,14 +89,18 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
     private ScrollbarWidget previewItemsScrollbar;
     @Nullable
     private ScrollbarWidget requestButtonsScrollbar;
+    @Nullable
+    private Button maxButton;
 
     private final List<AutocraftingRequestButton> requestButtons = new ArrayList<>();
     private final boolean requestsButtonsVisible;
 
     private final RateLimiter requestRateLimiter = RateLimiter.create(1);
+    private final RateLimiter maxAmountRequestRateLimiter = RateLimiter.create(1 / 5D);
 
     @Nullable
     private Double changedAmount;
+    private boolean mayEnableMaxAmountRequestButtonAgain;
 
     public AutocraftingPreviewScreen(final Screen parent,
                                      final Inventory playerInventory,
@@ -114,6 +145,7 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
 
     @Override
     protected void init() {
+        final boolean wasAlreadyInitialized = amountField != null;
         super.init();
         previewItemsScrollbar = new ScrollbarWidget(
             leftPos + 235,
@@ -128,13 +160,29 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
         if (confirmButton != null) {
             setStartDisabled();
         }
-        getMenu().loadCurrentRequest();
+        if (!wasAlreadyInitialized) {
+            getMenu().loadCurrentRequest();
+        }
+
+        final boolean wasActive = maxButton == null || maxButton.active;
+        maxButton = Button.builder(MAX, this::requestMaxAmount)
+            .size(22, 15)
+            .pos(leftPos + 185 - 1, topPos + 49 - 1)
+            .build();
+        maxButton.active = wasActive;
+        addRenderableWidget(maxButton);
+
         getExclusionZones().add(new Rect2i(
             leftPos - REQUESTS_WIDTH + 4,
             topPos,
             REQUESTS_WIDTH,
             REQUESTS_HEIGHT
         ));
+    }
+
+    private void requestMaxAmount(final Button button) {
+        button.active = false;
+        getMenu().requestMaxAmount();
     }
 
     private void initRequestButtons() {
@@ -254,16 +302,110 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
         final int x = leftPos + 8;
         final int y = topPos + 98;
         graphics.enableScissor(x, y, x + 221, y + PREVIEW_AREA_HEIGHT);
+        if (preview.type() == PreviewType.CYCLE_DETECTED) {
+            renderCycleDetected(graphics, y, x, preview);
+        } else if (preview.type() == PreviewType.OVERFLOW) {
+            renderRequestTooLargeToHandle(graphics, x, y);
+        } else {
+            renderPreviewRows(graphics, mouseX, mouseY, preview, y, x);
+        }
+        graphics.disableScissor();
+    }
+
+    private void renderCycleDetected(final GuiGraphics graphics, final int y, final int x, final Preview preview) {
+        int yy = y + 4;
+        SmallText.render(
+            graphics,
+            font,
+            CYCLE_DETECTED.getVisualOrderText(),
+            x + 4,
+            yy,
+            0xFF5555,
+            false
+        );
+        yy += 10;
+        SmallText.render(
+            graphics,
+            font,
+            CYCLE_OUTPUTS.getVisualOrderText(),
+            x + 4,
+            yy,
+            0x404040,
+            false
+        );
+        yy += 10;
+        for (final ResourceAmount output : preview.outputsOfPatternWithCycle()) {
+            final ResourceRendering rendering = RefinedStorageClientApi.INSTANCE.getResourceRendering(
+                output.resource().getClass()
+            );
+            rendering.render(output.resource(), graphics, x + 4, yy);
+            SmallText.render(
+                graphics,
+                font,
+                Component.literal(output.amount() + "x ").append(rendering.getDisplayName(output.resource()))
+                    .getVisualOrderText(),
+                x + 4 + 16 + 3,
+                yy + 5,
+                0x404040,
+                false
+            );
+            yy += 18;
+        }
+        yy += 2;
+        SmallText.render(
+            graphics,
+            font,
+            BREAK_THE_CYCLE_AND_TRY_AGAIN.getVisualOrderText(),
+            x + 4,
+            yy,
+            0x404040,
+            false
+        );
+    }
+
+    private void renderRequestTooLargeToHandle(final GuiGraphics graphics, final int x, final int y) {
+        SmallText.render(
+            graphics,
+            font,
+            REQUEST_TOO_LARGE_TO_HANDLE.getVisualOrderText(),
+            x + 4,
+            y + 4,
+            0xFF5555,
+            false
+        );
+        SmallText.render(
+            graphics,
+            font,
+            TRY_SMALLER_AMOUNT.getVisualOrderText(),
+            x + 4,
+            y + 4 + 10,
+            0x404040,
+            false
+        );
+    }
+
+    private void renderPreviewRows(final GuiGraphics graphics,
+                                   final int mouseX,
+                                   final int mouseY,
+                                   final Preview preview,
+                                   final int y,
+                                   final int x) {
         final List<PreviewItem> items = preview.items();
         final int rows = Math.ceilDiv(items.size(), COLUMNS);
         for (int i = 0; i < rows; ++i) {
-            final int scrollOffset = previewItemsScrollbar.isSmoothScrolling()
-                ? (int) previewItemsScrollbar.getOffset()
-                : (int) previewItemsScrollbar.getOffset() * ROW_HEIGHT;
+            final int scrollOffset = getScrollOffset();
             final int yy = y + (i * ROW_HEIGHT) - scrollOffset;
             renderRow(graphics, x, yy, i, items, mouseX, mouseY);
         }
-        graphics.disableScissor();
+    }
+
+    private int getScrollOffset() {
+        if (previewItemsScrollbar == null) {
+            return 0;
+        }
+        return (previewItemsScrollbar.isSmoothScrolling()
+            ? (int) previewItemsScrollbar.getOffset()
+            : (int) previewItemsScrollbar.getOffset() * ROW_HEIGHT);
     }
 
     private void renderRow(final GuiGraphics graphics,
@@ -338,6 +480,19 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
             0x404040,
             false
         );
+    }
+
+    @Override
+    protected void renderTooltip(final GuiGraphics graphics, final int x, final int y) {
+        super.renderTooltip(graphics, x, y);
+        if (maxButton != null && maxButton.isHovered()) {
+            Platform.INSTANCE.renderTooltip(
+                graphics,
+                List.of(HelpClientTooltipComponent.createAlwaysDisplayed(MAX_HELP)),
+                x,
+                y
+            );
+        }
     }
 
     @Override
@@ -424,6 +579,9 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
     }
 
     private void setPending() {
+        if (confirmButton == null) {
+            return;
+        }
         confirmButton.active = false;
         confirmButton.setError(false);
         confirmButton.setTooltip(null);
@@ -431,6 +589,9 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
     }
 
     private void setStartDisabled() {
+        if (confirmButton == null) {
+            return;
+        }
         confirmButton.active = false;
         confirmButton.setError(false);
         confirmButton.setTooltip(null);
@@ -443,6 +604,10 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
         if (changedAmount != null && requestRateLimiter.tryAcquire()) {
             getMenu().amountChanged(changedAmount);
             changedAmount = null;
+        }
+        if (mayEnableMaxAmountRequestButtonAgain && maxButton != null && maxAmountRequestRateLimiter.tryAcquire()) {
+            maxButton.active = true;
+            mayEnableMaxAmountRequestButtonAgain = false;
         }
     }
 
@@ -485,6 +650,13 @@ public class AutocraftingPreviewScreen extends AbstractAmountScreen<Autocrafting
             requestButton.setY(buttonY);
             requestButton.visible = isCraftingRequestButtonVisible(buttonY);
         }
+    }
+
+    @Override
+    public void maxAmountReceived(final double maxAmount) {
+        updateAmount(maxAmount);
+        maxAmountRequestRateLimiter.tryAcquire();
+        mayEnableMaxAmountRequestButtonAgain = true;
     }
 
     private void updateRequestsScrollbar() {